requirements.txt
Flask
gunicorn
tensorflow-cpu
numpy
pandas
